# refactor(fetch/Request): inline defaultInnerRequest