@AaronO 2021-09-27
> LGTM, can you do a quick benchmark for this too though?

Microbench similar to the old one:
```
❯ deno run t.js
inner_outer:         	n = 1000000, dt = 0.106s, r = 9433962/s, t = 106ns/op
inner_inline:        	n = 1000000, dt = 0.045s, r = 22222222/s, t = 45ns/op
inner_outer:         	n = 1000000, dt = 0.114s, r = 8771930/s, t = 114ns/op
inner_inline:        	n = 1000000, dt = 0.042s, r = 23809524/s, t = 42ns/op
inner_outer:         	n = 1000000, dt = 0.113s, r = 8849558/s, t = 113ns/op
inner_inline:        	n = 1000000, dt = 0.041s, r = 24390244/s, t = 41ns/op
```

In my profiled runs:
- `newInnerResponse`: 99.1ms -> 0.1ms (~1000x faster)
- `newInnerRequest`: 162.9ms -> 20.4ms (~8x faster), where virtually all of the remaining time is spent on `url.startsWith("blob:")`