@MierenManz 2021-09-27
> where virtually all of the remaining time is spent on `url.startsWith("blob:")`
Just out of curiosity. Is there any possible optimization for this or not really?